requirements.txt
mqtt-client
PyQt5==5.15.7
paho-mqtt==1.6.1
matplotlib==3.5.2
numpy==1.23.1